+++ src/vue/ui/feedback/index.ts
@@ -2,7 +2,7 @@ export { default as Feedback } from "./Feedback.vue"
 export { default as Pending } from "./Pending.vue"
 export { default as Success } from "./Success.vue"
 export { default as ErrorSummary } from "./ErrorSummary.vue"
-export { default as plugin } from "./plugin"
+export { default as plugin, useAppFeedback } from "./plugin"
 export {
     useFeedback,
     useFeedback as default,